+++ src/compiler/llvm_codegen_expr.c
@@ -33,7 +33,6 @@ static inline void llvm_emit_try_unwrap(GenContext *c, BEValue *value, Expr *exp
 static inline void llvm_emit_vector_initializer_list(GenContext *c, BEValue *value, Expr *expr);
 static inline void llvm_extract_bitvalue_from_array(GenContext *c, BEValue *be_value, Decl *member, Decl *parent_decl);
 static inline void llvm_emit_type_from_any(GenContext *c, BEValue *be_value);
-static inline void llvm_emit_memcmp(GenContext *c, BEValue *be_value, LLVMValueRef ptr, LLVMValueRef other_ptr, BinaryOp binary_op, AlignSize lhs_align, AlignSize rhs_align, ByteSize size);
 static void llvm_convert_vector_comparison(GenContext *c, BEValue *be_value, LLVMValueRef val, Type *vector_type,
 										   bool is_equals);
 static void llvm_emit_any_pointer(GenContext *c, BEValue *any, BEValue *pointer);
@@ -44,6 +43,7 @@ static void llvm_emit_initialize_designated_element(GenContext *c, BEValue *ref,
 static void llvm_emit_macro_body_expansion(GenContext *c, BEValue *value, Expr *body_expr);
 static void llvm_emit_post_unary_expr(GenContext *context, BEValue *be_value, Expr *expr);
 static void llvm_emit_unary_expr(GenContext *c, BEValue *value, Expr *expr);
+static inline void llvm_emit_memcmp(GenContext *c, BEValue *be_value, LLVMValueRef ptr, LLVMValueRef other_ptr, LLVMValueRef size);
 static LLVMTypeRef llvm_find_inner_struct_type_for_coerce(GenContext *c, LLVMTypeRef struct_type, ByteSize dest_size);
 static void llvm_expand_type_to_args(GenContext *context, Type *param_type, LLVMValueRef expand_ptr, LLVMValueRef *args, unsigned *arg_count_ref, AlignSize alignment);
 static inline void llvm_emit_initialize_reference_designated_bitstruct(GenContext *c, BEValue *ref, Decl *bitstruct, Expr **elements);
@@ -3564,8 +3564,8 @@ static void llvm_emit_struct_comparison(GenContext *c, BEValue *result, BEValue 
 	llvm_value_fold_optional(c, rhs);
 	llvm_value_addr(c, lhs);
 	llvm_value_addr(c, rhs);
-	llvm_emit_memcmp(c, result, lhs->value, rhs->value, binary_op, lhs->alignment, rhs->alignment,
-	                 type_size(lhs->type));
+	llvm_emit_memcmp(c, result, lhs->value, rhs->value, llvm_const_int(c, type_usz, type_size(lhs->type)));
+	llvm_emit_int_comp_zero(c, result, result, binary_op);
 }
 
 static inline LLVMValueRef llvm_emit_mult_int(GenContext *c, Type *type, LLVMValueRef left, LLVMValueRef right, SourceSpan loc)
@@ -3687,116 +3687,74 @@ INLINE bool should_inline_array_comp(ArraySize len, Type *base_type_lowered)
 	}
 }
 
-static void llvm_emit_memcmp_inline(GenContext *c, BEValue *be_value, LLVMValueRef lhs,
-                                    LLVMValueRef rhs, ByteSize element_size,
-									AlignSize lhs_align, AlignSize rhs_align, int len, bool want_match)
+static inline void llvm_emit_memcmp(GenContext *c, BEValue *be_value, LLVMValueRef ptr, LLVMValueRef other_ptr, LLVMValueRef size)
 {
-	assert(element_size <= platform_target.width_register / 8);
-	lhs_align = type_min_alignment(element_size, lhs_align);
-	rhs_align = type_min_alignment(element_size, rhs_align);
-	LLVMTypeRef element_type = LLVMIntTypeInContext(c->context, element_size * 8);
-	LLVMBasicBlockRef exit = llvm_basic_block_new(c, "array_cmp_exit");
-	LLVMBasicBlockRef loop_begin = llvm_basic_block_new(c, "array_loop_start");
-	LLVMBasicBlockRef comparison = llvm_basic_block_new(c, "array_loop_comparison");
-	LLVMBasicBlockRef comparison_phi;
-	LLVMBasicBlockRef loop_begin_phi;
-	LLVMValueRef len_val = llvm_const_int(c, type_usz, len);
-	LLVMValueRef one = llvm_const_int(c, type_usz, 1);
-	BEValue index_var;
-	llvm_value_set_address_abi_aligned(&index_var, llvm_emit_alloca_aligned(c, type_usz, "cmp.idx"), type_usz);
-	llvm_store_raw(c, &index_var, llvm_get_zero(c, type_usz));
-
-	llvm_emit_br(c, loop_begin);
-	llvm_emit_block(c, loop_begin);
-
-	AlignSize align_lhs;
-	BEValue lhs_v;
-	BEValue index_copy = index_var;
-	llvm_value_rvalue(c, &index_copy);
-
-	LLVMValueRef index_val = index_copy.value;
-	LLVMValueRef lhs_ptr = llvm_emit_pointer_inbounds_gep_raw(c, element_type, lhs, index_val);
-	LLVMValueRef rhs_ptr = llvm_emit_pointer_inbounds_gep_raw(c, element_type, rhs, index_val);
-	LLVMValueRef lhs_value = llvm_load(c, element_type, lhs_ptr, lhs_align, "lhs");
-	LLVMValueRef rhs_value = llvm_load(c, element_type, rhs_ptr, rhs_align, "rhs");
-	LLVMValueRef comp_val = LLVMBuildICmp(c->builder, LLVMIntEQ, lhs_value, rhs_value, "cmp");
-	loop_begin_phi = c->current_block;
-	llvm_emit_cond_br_raw(c, comp_val, comparison, exit);
-	llvm_emit_block(c, comparison);
-
-	LLVMValueRef new_index = LLVMBuildAdd(c->builder, index_copy.value, one, "inc");
-	llvm_store_raw(c, &index_var, new_index);
-	BEValue comp;
-	llvm_emit_int_comp_raw(c, &comp, type_usz, type_usz, new_index, len_val, BINARYOP_LT);
-	comparison_phi = c->current_block;
-	llvm_emit_cond_br(c, &comp, loop_begin, exit);
-	llvm_emit_block(c, exit);
-	LLVMValueRef success = LLVMConstInt(c->bool_type, want_match ? 1 : 0, false);
-	LLVMValueRef failure = LLVMConstInt(c->bool_type, want_match ? 0 : 1, false);
-	llvm_new_phi(c, be_value, "array_cmp_phi", type_bool, success, comparison_phi, failure, loop_begin_phi);
-
-}
-static void llvm_emit_memcmp_unrolled(GenContext *c, BEValue *be_value, LLVMValueRef lhs_ptr,
-									LLVMValueRef rhs_ptr, ByteSize element_size,
-									  AlignSize lhs_align, AlignSize rhs_align, int len, bool want_match)
-{
-	assert(len < 17);
-	assert(element_size <= platform_target.width_register / 8);
-	LLVMTypeRef element_type = LLVMIntTypeInContext(c->context, element_size * 8);
-	LLVMBasicBlockRef blocks[17];
-	LLVMValueRef value_block[17];
-	LLVMBasicBlockRef ok_block = llvm_basic_block_new(c, "match");
-	LLVMBasicBlockRef exit_block = llvm_basic_block_new(c, "exit");
-	LLVMValueRef success = LLVMConstInt(c->bool_type, want_match ? 1 : 0, false);
-	LLVMValueRef failure = LLVMConstInt(c->bool_type, want_match ? 0 : 1, false);
-	LLVMValueRef one = llvm_const_int(c, type_usz, 1);
-	for (unsigned i = 0; i < len; i++)
+	if (!c->memcmp_function)
 	{
-		value_block[i] = failure;
-		if (i > 0)
+		c->memcmp_function = LLVMGetNamedFunction(c->module, "memcmp");
+		if (!c->memcmp_function)
 		{
-			lhs_ptr = llvm_emit_pointer_inbounds_gep_raw(c, element_type, lhs_ptr, one);
-			rhs_ptr = llvm_emit_pointer_inbounds_gep_raw(c, element_type, rhs_ptr, one);
+			c->memcmp_function = LLVMAddFunction(c->module, "memcmp", c->memcmp_function_type);
 		}
-		AlignSize lhs_align_current = type_min_alignment(lhs_align + i * element_size, lhs_align);
-		AlignSize rhs_align_current = type_min_alignment(rhs_align + i * element_size, rhs_align);
-		LLVMValueRef lhs_value = llvm_load(c, element_type, lhs_ptr, lhs_align_current, "lhs");
-		LLVMValueRef rhs_value = llvm_load(c, element_type, rhs_ptr, rhs_align_current, "rhs");
-		LLVMValueRef comp = LLVMBuildICmp(c->builder, LLVMIntEQ, lhs_value, rhs_value, "cmp");
-		blocks[i] = c->current_block;
-		LLVMBasicBlockRef block = ok_block;
-		block = i < len - 1 ? llvm_basic_block_new(c, "next_check") : block;
-		llvm_emit_cond_br_raw(c, comp, block, exit_block);
-		llvm_emit_block(c, block);
 	}
-	llvm_emit_br(c, exit_block);
-	llvm_emit_block(c, exit_block);
-	value_block[len] = success;
-	blocks[len] = ok_block;
-	LLVMValueRef phi = LLVMBuildPhi(c->builder, c->bool_type, "memcmp_phi");
-	LLVMAddIncoming(phi, value_block, blocks, len + 1);
-	llvm_value_set(be_value, phi, type_bool);
+	LLVMValueRef args[3] = { ptr, other_ptr, size };
+	LLVMValueRef function = LLVMBuildCall2(c->builder, c->memcmp_function_type, c->memcmp_function, args, 3, "cmp");
+	llvm_value_set(be_value, function, type_cint);
 }
-static inline void llvm_emit_memcmp(GenContext *c, BEValue *be_value, LLVMValueRef ptr, LLVMValueRef other_ptr, BinaryOp binary_op, AlignSize lhs_align, AlignSize rhs_align, ByteSize size)
-{
 
-	ByteSize element_size = lhs_align > platform_target.width_register / 8 ? platform_target.width_register / 8 : lhs_align;
-	if (element_size > rhs_align) element_size = rhs_align;
-	if (element_size > size) element_size = size;
-	ByteSize repeats = size / element_size;
-	assert(size % element_size == 0 && "Expected size padded to alignment");
-	if (repeats <= MEMCMP_INLINE_REGS)
-	{
-		llvm_emit_memcmp_unrolled(c, be_value, ptr, other_ptr, element_size, lhs_align, rhs_align, repeats, binary_op == BINARYOP_EQ);
-		return;
-	}
-	llvm_emit_memcmp_inline(c, be_value, ptr, other_ptr, element_size, lhs_align, rhs_align, repeats, binary_op == BINARYOP_EQ);
-}
 static void llvm_emit_array_comp(GenContext *c, BEValue *be_value, BEValue *lhs, BEValue *rhs, BinaryOp binary_op)
 {
+	Type *array_base = type_flatten(lhs->type->array.base);
+	switch (array_base->type_kind)
+	{
+
+		case ALL_INTS:
+		case TYPE_POINTER:
+		case TYPE_ENUM:
+		case TYPE_FUNC_PTR:
+		case TYPE_INTERFACE:
+		case TYPE_ANY:
+		case TYPE_ANYFAULT:
+		case TYPE_FAULTTYPE:
+		case TYPE_TYPEID:
+MEMCMP:
+			llvm_value_addr(c, lhs);
+			llvm_value_addr(c, rhs);
+			llvm_emit_memcmp(c, be_value, lhs->value, rhs->value, llvm_const_int(c, type_usz, type_size(lhs->type)));
+			llvm_emit_int_comp_zero(c, be_value, be_value, binary_op);
+			return;
+		case TYPE_VECTOR:
+			if (is_power_of_two(array_base->array.len)) goto MEMCMP;
+			break;
+		case TYPE_UNION:
+		case TYPE_STRUCT:
+		case TYPE_BITSTRUCT:
+			if (array_base->decl->attr_compact) goto MEMCMP;
+			break;
+		case TYPE_POISONED:
+		case TYPE_VOID:
+		case TYPE_DISTINCT:
+		case TYPE_FUNC_RAW:
+		case TYPE_TYPEDEF:
+		case TYPE_INFERRED_ARRAY:
+		case TYPE_INFERRED_VECTOR:
+		case TYPE_UNTYPED_LIST:
+		case TYPE_OPTIONAL:
+		case TYPE_WILDCARD:
+		case TYPE_TYPEINFO:
+		case TYPE_MEMBER:
+			UNREACHABLE
+		case ALL_FLOATS:
+		case TYPE_SLICE:
+		case TYPE_ARRAY:
+		case TYPE_FLEXIBLE_ARRAY:
+		case TYPE_BOOL:
+			break;
+	}
+
 	bool want_match = binary_op == BINARYOP_EQ;
 	ArraySize len = lhs->type->array.len;
-	Type *array_base_type = type_lowering(lhs->type->array.base);
+	Type *array_base_type = type_lowering(array_base);
 	LLVMTypeRef array_type = llvm_get_type(c, lhs->type);
 	if (should_inline_array_comp(len, array_base_type))
 	{